+++ vspec/vssexporters/vss2ddsidl.py
